add Categories() and update Clienter
